[hardware] :bug: Handle hazards globally from main sequencer
When the sequencer issues an instruction, it also saves the information
about its hazards in an entry of the global_hazard_table.
This table keeps, for each instruction, up to date information on the
status of the offender previous instructions. This information is given
to the operand requesters, which can handle the stalls downstream.
Before the introduction of the global hazard table, each instruction
kept, in the pipeline, information about the status of the offending
instructions. When this is the case, these information must be kept
up to date in multiple points of the pipeline.
With the introduction of new fall-through registers to cut valid-ready
inout paths, updating the hazard bits within the pipeline was becoming
more and more cumbersome. The global hazard table should solve this
problem in a more elegan way.